+++ crates/ryo3-jiff/src/constants.rs
@@ -1,0 +1,4 @@
+use jiff::fmt::temporal::{DateTimeParser, SpanParser};
+
+pub(crate) static DATETIME_PARSER: DateTimeParser = DateTimeParser::new();
+pub(crate) static SPAN_PARSER: SpanParser = SpanParser::new();